fixed updating from paid event to not paid event and vice versa

+++ app/payment/tasks.py
@@ -3,13 +3,15 @@ from app.content.models.registration import Registration
 from app.payment.models.order import Order
 from app.payment.enums import OrderStatus
 from app.util.tasks import BaseTask
+from app.payment.views.vipps_callback import vipps_callback
 
 @app.task(bind=True, base=BaseTask)
 def check_if_has_paid(self, order_id, registration_id):
     try:
+        vipps_callback(None, order_id)
         order = Order.objects.get(order_id=order_id)
         order_status = order.status
-        if order_status != OrderStatus.CAPTURE and order_status != OrderStatus.RESERVE:
+        if order_status != OrderStatus.CAPTURE and order_status != OrderStatus.RESERVE and order_status != OrderStatus.SALE:
             Registration.objects.filter(registration_id=registration_id).delete()
 
     except Exception as e:
